Rewriting the code to use folders.
