+++ DEA/Graph/GraphHelper.cs
@@ -238,144 +238,6 @@ namespace DEA
         }
 
         /// <summary>
-        /// Check if the folder paths and folders exsits. If not they are created this is initiated on program launch.
-        /// </summary>
-        /// <param name="folderSwitch"></param>
-        /// <returns>Folder paths have use the correct download folder name.</returns>
-        public static string CheckFolders(string folderSwitch)
-        {
-            // Get current execution path.
-            string folderPath = string.Empty;
-            string? pthRootFolder = Path.GetDirectoryName(System.Reflection.Assembly.GetExecutingAssembly().Location);
-            string downloadFolderName = "Download";
-            string attachmentFolderName = "Attachments";
-            string ftpFileFolderName = "FTPFiles";
-            string logFolderName = "Logs";
-            string pathDownloadFolder = Path.Combine(pthRootFolder!, downloadFolderName); // Main file download folder.
-            string pathLogFolder = Path.Combine(pthRootFolder!, logFolderName); // Log folder
-            string attachmentPath = Path.Combine(pathDownloadFolder!, attachmentFolderName); // Email attachment download folder.
-            string ftpPath = Path.Combine(pathDownloadFolder, ftpFileFolderName); // FTP file download folder.
-
-            // Check if download folder exists. If not creates the fodler.
-            if (!System.IO.Directory.Exists(pathDownloadFolder))
-            {
-                try
-                {
-                    System.IO.Directory.CreateDirectory(pathDownloadFolder);
-                }
-                catch (Exception ex)
-                {
-                    WriteLogClass.WriteToLog(1, $"Exception at download folder creation: {ex.Message}", 1);
-                }
-            }
-
-            if (!System.IO.Directory.Exists(attachmentPath))
-            {
-                try
-                {
-                    System.IO.Directory.CreateDirectory(attachmentPath);
-                }
-                catch (Exception ex)
-                {
-                    WriteLogClass.WriteToLog(1, $"Exception at attachmnet folder creation: {ex.Message}", 1);
-                }
-            }
-
-            if (System.IO.Directory.Exists(ftpPath))
-            {
-                try
-                {
-                    System.IO.Directory.CreateDirectory(ftpPath);
-                }
-                catch (Exception ex)
-                {
-                    WriteLogClass.WriteToLog(1, $"Exception at FTP folder creation: {ex.Message}", 1);
-                }
-            }
-
-            if (!System.IO.Directory.Exists(pathLogFolder))
-            {
-                try
-                {
-                    System.IO.Directory.CreateDirectory(pathLogFolder);
-                }
-                catch (Exception ex)
-                {
-                    WriteLogClass.WriteToLog(1, $"Exception at download folder creation: {ex.Message}", 1);
-                }
-            }
-
-            if (folderSwitch.ToLower() == "download")
-            {
-                folderPath = pathDownloadFolder;
-            }
-            else if (folderSwitch.ToLower() == "log")
-            {
-                folderPath = pathLogFolder;
-            }
-            else if (folderSwitch.ToLower() == "email")
-            {
-                folderPath = attachmentPath;
-            }
-            else if (folderSwitch.ToLower() == "ftp")
-            {
-                folderPath = ftpPath;
-            }
-            else
-            {
-                folderPath = string.Empty;
-            }
-
-            return folderPath;
-        }
-
-        /// <summary>
-        /// Downnloads the attachment files to the local harddrive.
-        /// </summary>
-        /// <param name="DownloadFolderPath"></param>
-        /// <param name="DownloadFileName"></param>
-        /// <param name="DownloadFileData"></param>
-        /// <returns></returns>
-        public static bool DownloadAttachedFiles(string DownloadFolderPath, string DownloadFileName, byte[] DownloadFileData)
-        {
-            if (!System.IO.Directory.Exists(DownloadFolderPath))
-            {
-                try
-                {
-                    System.IO.Directory.CreateDirectory(DownloadFolderPath);
-                }
-                catch (Exception ex)
-                {
-                    WriteLogClass.WriteToLog(1, $"Exception at download folder creation: {ex.Message}", 1);
-                }
-            }
-
-            try
-            {
-                var PathFullDownloadFile = Path.Combine(DownloadFolderPath, DownloadFileName);
-                var FileNameOnly = Path.GetFileNameWithoutExtension(PathFullDownloadFile);
-                var FileExtention = Path.GetExtension(PathFullDownloadFile);
-                var FilePathOnly = Path.GetDirectoryName(PathFullDownloadFile);
-                int Count = 1;
-
-                while (System.IO.File.Exists(PathFullDownloadFile)) // If file exists starts to rename from next file.
-                {
-                    var NewFileName = string.Format("{0}({1})", FileNameOnly, Count++); // Makes the new file name.
-                    PathFullDownloadFile = Path.Combine(FilePathOnly!, NewFileName + FileExtention); // Set tthe new path as the download file path.
-                }
-
-                // Full path for the attachment to be downloaded with the attachment name                
-                System.IO.File.WriteAllBytes(PathFullDownloadFile, DownloadFileData);
-                return true;
-            }
-            catch (Exception ex)
-            {
-                WriteLogClass.WriteToLog(1, $"Exception at download path: {ex.Message}", 1);
-                return false;
-            }
-        }
-
-        /// <summary>
         /// Moves the email to Downloded folder.
         /// </summary>
         /// <param name="FirstFolderId"></param>
